Fix logic in calculator in first run & after
When 'equals' is clicked when there are no numbers, the total
will become undefined and then when clicking on operand with number
it will return with NaN. For example: = is clicked then + 1 is clicked
it will display NaN. Added logic to allow previous answer to be
continuously worked on. Lastly, when new number were clicked after
an answer, it would add up to the previous answer instead of starting
a new one.

Creating an if statement when there is an operand and an operator
is clicked, to check if there isn't a number it will make the previous
answer equal to num1 and take in the users next number. Creating a
condition when = is clicked to check if there is a number in num1 and
num2 to prevent NaN. Added an additional condition to allow num2
be ran more loosely.